+++ magame/source/LevelChoiceScreen.h
@@ -75,8 +75,8 @@ private:
     SpritePtr m_entry_disabled_sprite;   // animacja nieaktywnego wejścia
     SpritePtr m_face_sprite;             // animacja sprite'a bohatera
 
-    float m_tile_width;             // szerokość kafla na ekranie
-    float m_tile_height;            // wysokość kafla na ekranie
+    double m_tile_width;             // szerokość kafla na ekranie
+    double m_tile_height;            // wysokość kafla na ekranie
 
     PlayerPtr m_player;             // wskaźnik na gracza, który przekażemy do instancji gry
     boost::shared_ptr<AppState> m_next_app_state;
